refactor(plugin-sdk): move processor descriptor types to streamlib-processor-schema
Phase 2a step 1. ProcessorDescriptor / PortDescriptor / ConfigField /
ConfigDescriptor / CodeExamples / ProcessorRuntime + the port_schema_spec_wire
serde mirror are engine-free serde data (deps: SchemaIdent / PortSchemaSpec /
ProcessorScheduling from processor-schema + serde). They cross the cdylib
plugin ABI as rmp_serde msgpack, so a single definition in the shared
engine-free crate keeps host and plugin byte-identical. The engine re-exports
them at core::descriptors unchanged; the round-trip regression tests move with
the types. Workspace cargo check green; the 4 msgpack round-trip tests pass.
